Refactor obsolete code
